+++ src/MyTypes.cpp
@@ -35,7 +35,7 @@ Annotation::Annotation(const unsigned int natoms,
 
 /* ------------------------------------------------------------------------- */
 
-bool Annotation::containsAtom(const int atom_id) const {
+bool Annotation::containsAtom(const long atom_id) const {
 
 	unsigned int i;
 	for (i = 0; i < natoms; i++) {
@@ -49,7 +49,7 @@ bool Annotation::containsAtom(const int atom_id) const {
 
 /* ------------------------------------------------------------------------- */
 
-void Annotation::adjustIds(const int atom_id, int offset) {
+void Annotation::adjustIds(const long atom_id, int offset) {
 
 	unsigned int i;
 	for (i = 0; i < natoms; i++) {
@@ -60,7 +60,7 @@ void Annotation::adjustIds(const int atom_id, int offset) {
 
 /* ------------------------------------------------------------------------- */
 
-int Annotation::getAtom(const int index) const {
+long Annotation::getAtom(const long index) const {
 
 	if (index < 0 || index >= natoms) return -1;
 	return atoms[index];
