+++ src/myFiles.h
@@ -16,7 +16,7 @@ void OpenFile(void);
 BufferFile * OpenDatFile(void);
 long LocateKeyWord(const char *Buffer, const char * KeyWord, long length, long bytecount);
 long ParseGLogLine(BufferFile * Buffer, Frame * lFrame, long numExpected, long Type, float *maxsize);
-long ParseCartLine(char *Line, long *atomtype, CPoint3D *coord, CPoint3D *offset, long Mode);
+int ParseCartLine(char *Line, int *atomtype, CPoint3D *coord, CPoint3D *offset, long Mode);
 int SetAtomType(const unsigned char *TestLabel);
 
 #endif
